Add common/edge.c
