+++ include/common/direction.h
@@ -3,7 +3,7 @@
 #define LABWC_DIRECTION_H
 
 #include <wlr/types/wlr_output_layout.h>
-#include "config/types.h"
+#include "common/edge.h"
 
 bool direction_from_edge(enum lab_edge edge, enum wlr_direction *direction);
 enum wlr_direction direction_get_opposite(enum wlr_direction direction);
